requirements.txt
lnetatmo
suntime
Pillow